Removed tolerance, default threshold, and new icon
TODO:
Modify is_inside_path to be more reliable. Make it so when counting
intersections, it skips over contiguous edge pixels since that should
count as a single intersection.
Modify the find_edge logic to look look as far as 2x the distance it
normally stops at IF the user first clicks on a point that's not within
threshold, so the user could click in the center of the bone and still
get accurate results.