Changed way it finds best_path
Known issues:
As you get cloesr to the knee, the density in the center of the bone
becomes less uniform, so a bunch of really small paths are detected even
at low threshold values.
May want to modify find_edges to search for ALL edge points within a set
distance, rather than get just the first two.
Also, intuitively, theoretically there's no reason to store or even
investigate the first edge point found in a given direction once a
subsequent one is found, as the last point found in a given direction
should be guaranteed to contain the path of the inner edge point.